+++ provider/datasource_rediscloud_active_active_subscription.go
@@ -64,6 +64,11 @@ func dataSourceRedisCloudActiveActiveSubscription() *schema.Resource {
 				Type:        schema.TypeBool,
 				Computed:    true,
 			},
+			"prometheus_endpoint": {
+				Description: "The Prometheus scrape endpoint for databases in this subscription",
+				Type:        schema.TypeString,
+				Computed:    true,
+			},
 			"cloud_provider": {
 				Description: "A cloud provider string either GCP or AWS",
 				Type:        schema.TypeString,
@@ -264,6 +269,14 @@ func dataSourceRedisCloudActiveActiveSubscriptionRead(ctx context.Context, d *sc
 		publicEndpointAccess = redis.BoolValue(sub.PublicEndpointAccess)
 	}
 	if err := d.Set("public_endpoint_access", publicEndpointAccess); err != nil {
+		return diag.FromErr(err)
+	}
+
+	prometheusEndpoint := ""
+	if sub.PrometheusEndpoint != nil {
+		prometheusEndpoint = redis.StringValue(sub.PrometheusEndpoint)
+	}
+	if err := d.Set("prometheus_endpoint", prometheusEndpoint); err != nil {
 		return diag.FromErr(err)
 	}