+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 See updating [Changelog example here](https://keepachangelog.com/en/1.0.0/)
 
 
+# Unreleased
+
+## Changed
+- CI: Added VCR support for recording and replaying API interactions in acceptance tests, with the passwordless Pro database tests as the first to use it.
+
+
 # 2.15.0 (28th April 2026)
 
 ## Added