+++ CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See updating [Changelog example here](https://keepachangelog.com/en/1.0.0/)
 
+# Unreleased
+
+## Fixed
+- `rediscloud_active_active_subscription_database`: Changing the database `name` attribute no longer causes the resource to be destroyed and recreated.
 
 # 2.15.0 (28th April 2026)