+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 All notable changes to this project will be documented in this file.
 See updating [Changelog example here](https://keepachangelog.com/en/1.0.0/)
 
-# Unreleased
+# 2.16.0 (May 2026)
 
 ## Added
 - Added `prometheus_endpoint` attribute to the `rediscloud_subscription` resource.